+++ packages/main-api/src/models/dynamo_tables/main/email/mod.rs
@@ -1,0 +1,3 @@
+pub mod email_verification;
+
+pub use email_verification::*;
