+++ packages/main-api/src/models/dynamo_tables/main/mod.rs
@@ -1,3 +1,3 @@
-pub mod email_verification;
-pub mod feeds;
+pub mod email;
+pub mod feed;
 pub mod user;
