+++ packages/main-api/src/models/dynamo_tables/main/user/mod.rs
@@ -1,0 +1,3 @@
+pub mod user;
+
+pub use user::*;
